pyproject.toml
[tool.poetry]
name = "prism"
version = "0.1.0"
description = "Platform for Resource Intelligence & Subsurface Management"
authors = ["PRISM Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14"
django-filter = "^23.0"
mozilla-django-oidc = "^4.0"
python-jose = {extras = ["cryptography"], version = "^3.3"}
psycopg = {extras = ["binary"], version = "^3.1"}
redis = "^5.0"
celery = "^5.3"
django-redis = "^5.4"
django-environ = "^0.11"
whitenoise = "^6.6"
gunicorn = "^21.2"
drf-spectacular = "^0.27"
drf-spectacular-sidecar = "^2024.1"
django-htmx = "^1.17"
django-prometheus = "^2.3"
pandas = "^2.1"
numpy = "^1.26"
scipy = "^1.11"
jsonschema = "^4.20"
semver = "^3.0"
crispy-bootstrap5 = "^2024.2"
django-crispy-forms = "^2.1"
paramiko = "^3.4"
reportlab = "^4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
pytest-cov = "^4.1"
black = "^23.12"
isort = "^5.13"
flake8 = "^7.0"
mypy = "^1.7"
types-requests = "^2.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "prism_site.settings"
python_files = ["test_*.py", "*_test.py"]